+++ docs/release-notes.md
@@ -1,6 +1,15 @@
 ← [README](README.md)
 
 # Release notes
+## 4.0.6
+Released 07 April 2024 for Stardew Valley 1.6.0 or later.
+
+* For player:
+  * The SMAPI log file now includes installed mod IDs, to help with troubleshooting (thanks to DecidedlyHuman!).
+
+* For mod authors:
+  * Added optional [`MinimumGameVersion` manifest field](https://stardewvalleywiki.com/Modding:Modder_Guide/APIs/Manifest#Minimum_SMAPI_or_game_version).
+
 ## 4.0.5
 Released 06 April 2024 for Stardew Valley 1.6.0 or later.